Update for Swift 5.0
