+++ Package.swift
@@ -1,9 +1,21 @@
-
+// swift-tools-version:5.0
+// The swift-tools-version declares the minimum version of Swift required to
+// build this package.
 import PackageDescription
 
 let package = Package(
-    name: "Axiomatic",
-    dependencies: [
-        .Package(url: "https://github.com/JadenGeller/Gluey.git", majorVersion: 1)
-    ]
+  name: "Axiomatic",
+  dependencies: [
+    .package(url: "https://github.com/MaxDesiatov/Gluey.git", .branch("master")),
+  ],
+  targets: [
+    // Targets are the basic building blocks of a package. A target can define
+    // a module or a test suite.
+    // Targets can depend on other targets in this package, and on products
+    // in packages which this package depends on.
+    .target(
+      name: "Axiomatic",
+      dependencies: ["Gluey"]
+    ),
+  ]
 )
